+++ components/openai/OpenAI.c
@@ -128,7 +128,9 @@ typedef struct {
 
     char *(*get)(const char *base_url, const char *api_key, const char *endpoint);                                                     /*!<  Perform an HTTP GET request. */
     char *(*del)(const char *base_url, const char *api_key, const char *endpoint);                                                     /*!<  Perform an HTTP DELETE request. */
-    char *(*post)(const char *base_url, const char *api_key, const char *endpoint, char *jsonBody,  size_t *output_len);               /*!<  Perform an HTTP POST request. */
+    char *(*post)(const char *base_url, const char *api_key, const char *endpoint, char *jsonBody, size_t *output_len);                /*!<  Perform an HTTP POST request. */
+    char *(*speechpost)(const char *base_url, const char *api_key, const char *endpoint, char *jsonBody, size_t *output_len);          /*!<  Perform an HTTP POST request for speech. */
+    char *(*speechpost_stream)(const char *base_url, const char *api_key, const char *endpoint, char *jsonBody, size_t *output_len, OpenAI_StreamCallback stream_callback);          /*!<  Perform an HTTP POST request for stream speech. */
     char *(*upload)(const char *base_url, const char *api_key, const char *endpoint, const char *boundary, uint8_t *data, size_t len); /*!<  Upload data using an HTTP request. */
 } _OpenAI_t;
 
@@ -1056,35 +1058,138 @@ static void OpenAI_ChatCompletionClearConversation(OpenAI_ChatCompletion_t *chat
     }
 }
 
-static cJSON *createChatMessage(cJSON *messages, const char *role, const char *content)
+static cJSON *createContentObject(const char *type, const char *value)
+{
+    OPENAI_ERROR_CHECK(type != NULL, "type is NULL!", NULL);
+    OPENAI_ERROR_CHECK(value != NULL, "value is NULL!", NULL);
+
+    cJSON *content_obj = cJSON_CreateObject();
+    OPENAI_ERROR_CHECK(content_obj != NULL, "Failed to create content_obj!", NULL);
+
+    OPENAI_ERROR_CHECK_GOTO(cJSON_AddStringToObject(content_obj, "type", type), "Failed to add 'type' field!", cleanup);
+
+    if (strcmp(type, "text") == 0) {
+        OPENAI_ERROR_CHECK_GOTO(cJSON_AddStringToObject(content_obj, "text", value), "Failed to add 'text' field!", cleanup);
+    } else if (strcmp(type, "image_url") == 0) {
+        cJSON *image_url_obj = cJSON_CreateObject();
+        OPENAI_ERROR_CHECK_GOTO(image_url_obj != NULL, "Failed to create image_url_obj!", cleanup);
+        if (cJSON_AddStringToObject(image_url_obj, "url", value) == NULL) {
+            cJSON_Delete(image_url_obj);
+            ESP_LOGE(TAG, "Failed to add 'url' field!");
+            goto cleanup;
+        }
+        if (!cJSON_AddItemToObject(content_obj, "image_url", image_url_obj)) {
+            cJSON_Delete(image_url_obj);
+            ESP_LOGE(TAG, "Failed to add 'image_url' field!");
+            goto cleanup;
+        }
+    } else if (strcmp(type, "input_audio") == 0) {
+        /*
+         * Because we don't provide an interface to specify the audio format,
+         * we use the file's magic number to determine the audio format.
+         *
+         * OpenAI only supports mp3 and wav formats. 0 for unknown, 1 for mp3, 2 for wav.
+         */
+        uint8_t audio_format = 0;
+        const size_t audio_size = strlen(value);
+
+        OPENAI_ERROR_CHECK_GOTO(audio_size > 16, "Audio file is too small!", cleanup);
+        /*
+         * For WAV files, the magic number is string `RIFF____WAVE`, _ means any character.
+         * After base64 encoding, the magic string is `UklG_______QVZF`, _ means any character.
+         *
+         * For MP3 files. the magic number is string `ID3`, hex `0xFFFB`, `0xFFF3` or `0xFFF2`.
+         * After base64 encoding, the magic string is `SUQz` or `//`
+         */
+        if (memcmp(value, "UklG", 4) == 0 && memcmp(value + 12, "QVZF", 4) == 0) {
+            audio_format = 2;
+        } else if (memcmp(value, "SUQz", 4) == 0 || memcmp(value, "//", 2) == 0) {
+            audio_format = 1;
+        }
+        OPENAI_ERROR_CHECK_GOTO(audio_format != 0, "Unknown audio format!", cleanup);
+
+        cJSON *audio_obj = cJSON_CreateObject();
+        OPENAI_ERROR_CHECK_GOTO(audio_obj, "Failed to create audio_obj!", cleanup);
+
+        if (!cJSON_AddStringToObject(audio_obj, "format", audio_format == 1 ? "mp3" : "wav")) {
+            cJSON_Delete(audio_obj);
+            ESP_LOGE(TAG, "Failed to add format to audio_obj!");
+            goto cleanup;
+        }
+        if (!cJSON_AddStringToObject(audio_obj, "data", value)) {
+            cJSON_Delete(audio_obj);
+            ESP_LOGE(TAG, "Failed to add data to audio_obj!");
+            goto cleanup;
+        }
+        if (!cJSON_AddItemToObject(content_obj, "input_audio", audio_obj)) {
+            cJSON_Delete(audio_obj);
+            ESP_LOGE(TAG, "Failed to add audio_obj to content_obj!");
+            goto cleanup;
+        }
+    } else {
+        ESP_LOGW(TAG, "Unknown type: %s, skip building extra fields", type);
+    }
+
+    return content_obj;
+
+cleanup:
+    if (content_obj) {
+        cJSON_Delete(content_obj);
+    }
+    return NULL;
+}
+
+static cJSON *createMultiModalChatMessage(const char *role, const char *type, const char *value)
 {
     cJSON *message = cJSON_CreateObject();
-    OPENAI_ERROR_CHECK(message != NULL, "cJSON_CreateObject failed!", NULL);
+    OPENAI_ERROR_CHECK(message != NULL, "Failed to create message object!", NULL);
     if (cJSON_AddStringToObject(message, "role", role) == NULL) {
         cJSON_Delete(message);
-        ESP_LOGE(TAG, "cJSON_AddStringToObject failed!");
+        ESP_LOGE(TAG, "Failed to add role field!");
         return NULL;
     }
-    if (cJSON_AddStringToObject(message, "content", content) == NULL) {
+
+    cJSON *content_arr = cJSON_CreateArray();
+    if (!content_arr) {
         cJSON_Delete(message);
-        ESP_LOGE(TAG, "cJSON_AddStringToObject failed!");
+        ESP_LOGE(TAG, "Failed to create content array!");
         return NULL;
     }
-    if (!cJSON_AddItemToArray(messages, message)) {
+
+    cJSON *content_obj = createContentObject(type, value);
+    if (!content_obj) {
         cJSON_Delete(message);
-        ESP_LOGE(TAG, "cJSON_AddItemToArray failed!");
+        cJSON_Delete(content_arr);
         return NULL;
     }
+
+    if (!cJSON_AddItemToArray(content_arr, content_obj)) {
+        cJSON_Delete(message);
+        cJSON_Delete(content_arr);
+        cJSON_Delete(content_obj);
+        ESP_LOGE(TAG, "Failed to add content_obj to content array!");
+        return NULL;
+    }
+
+    if (!cJSON_AddItemToObject(message, "content", content_arr)) {
+        cJSON_Delete(message);
+        cJSON_Delete(content_arr);
+        ESP_LOGE(TAG, "Failed to add content array to message!");
+        return NULL;
+    }
+
     return message;
 }
 
-OpenAI_StringResponse_t *OpenAI_ChatCompletionMessage(OpenAI_ChatCompletion_t *chatCompletion, const char *p, bool save)
+OpenAI_StringResponse_t *OpenAI_ChatCompletionMultiModalMessage(OpenAI_ChatCompletion_t *chatCompletion, const char *type, const char *contentValue, bool save)
 {
+    const char *role = "user";
     const char *endpoint = "chat/completions";
     OpenAI_StringResponse_t *result = NULL;
 
     cJSON *req = cJSON_CreateObject();
     OPENAI_ERROR_CHECK(req != NULL, "cJSON_CreateObject failed!", result);
+
     _OpenAI_ChatCompletion_t *_chatCompletion = __containerof(chatCompletion, _OpenAI_ChatCompletion_t, parent);
     reqAddString("model", (_chatCompletion->model == NULL) ? "gpt-3.5-turbo" : _chatCompletion->model);
 
@@ -1095,11 +1200,19 @@ OpenAI_StringResponse_t *OpenAI_ChatCompletionMessage(OpenAI_ChatCompletion_t *c
         ESP_LOGE(TAG, "cJSON_CreateArray failed!");
         return result;
     }
-    if (_chatCompletion->description != NULL) {
-        if (createChatMessage(_messages, "system", _chatCompletion->description) == NULL) {
+    if (_chatCompletion->description) {
+        cJSON *system_msg = createMultiModalChatMessage("system", "text", _chatCompletion->description);
+        if (!system_msg) {
             cJSON_Delete(req);
             cJSON_Delete(_messages);
-            ESP_LOGE(TAG, "createChatMessage failed!");
+            ESP_LOGE(TAG, "Failed to create system_msg!");
+            return result;
+        }
+        if (!cJSON_AddItemToArray(_messages, system_msg)) {
+            cJSON_Delete(req);
+            cJSON_Delete(_messages);
+            cJSON_Delete(system_msg);
+            ESP_LOGE(TAG, "Failed to add system_msg!");
             return result;
         }
     }
@@ -1117,10 +1230,18 @@ OpenAI_StringResponse_t *OpenAI_ChatCompletionMessage(OpenAI_ChatCompletion_t *c
             }
         }
     }
-    if (createChatMessage(_messages, "user", p) == NULL) {
+    cJSON *new_msg = createMultiModalChatMessage(role, type, contentValue);
+    if (!new_msg) {
         cJSON_Delete(req);
         cJSON_Delete(_messages);
-        ESP_LOGE(TAG, "createChatMessage failed!");
+        ESP_LOGE(TAG, "Failed to create new_msg!");
+        return result;
+    }
+    if (!cJSON_AddItemToArray(_messages, new_msg)) {
+        cJSON_Delete(req);
+        cJSON_Delete(_messages);
+        cJSON_Delete(new_msg);
+        ESP_LOGE(TAG, "Failed to add new_msg!");
         return result;
     }
 
@@ -1147,6 +1268,7 @@ OpenAI_StringResponse_t *OpenAI_ChatCompletionMessage(OpenAI_ChatCompletion_t *c
         reqAddString("user", _chatCompletion->user);
     }
     char *jsonBody = cJSON_Print(req);
+
     cJSON_Delete(req);
     char *res = _chatCompletion->oai->post(_chatCompletion->oai->base_url, _chatCompletion->oai->api_key, endpoint, jsonBody, NULL);
     free(jsonBody);
@@ -1155,12 +1277,13 @@ OpenAI_StringResponse_t *OpenAI_ChatCompletionMessage(OpenAI_ChatCompletion_t *c
         //add the responses to the messages here
         //double parsing is here as workaround
         OpenAI_StringResponse_t *r = OpenAI_StringResponseCreate(res);
-        if (r->getLen(r)) {
-            if (createChatMessage(_chatCompletion->messages, "user", p) == NULL) {
-                ESP_LOGE(TAG, "createChatMessage failed!");
-            }
-            if (createChatMessage(_chatCompletion->messages, "assistant", r->getData(r, 0)) == NULL) {
-                ESP_LOGE(TAG, "createChatMessage failed!");
+        if (r && r->getLen(r)) {
+            const char *assistant_text = r->getData(r, 0);
+            cJSON *assistant_msg = createMultiModalChatMessage("assistant", "text", assistant_text);
+            if (assistant_msg) {
+                cJSON_AddItemToArray(_chatCompletion->messages, assistant_msg);
+            } else {
+                ESP_LOGE(TAG, "Failed to create assistant_msg!");
             }
         }
         free(res);
@@ -1169,6 +1292,11 @@ OpenAI_StringResponse_t *OpenAI_ChatCompletionMessage(OpenAI_ChatCompletion_t *c
     result = OpenAI_StringResponseCreate(res);
     free(res);
     return result;
+}
+
+OpenAI_StringResponse_t *OpenAI_ChatCompletionMessage(OpenAI_ChatCompletion_t *chatCompletion, const char *contentValue, bool save)
+{
+    return OpenAI_ChatCompletionMultiModalMessage(chatCompletion, "text", contentValue, save);
 }
 
 static OpenAI_ChatCompletion_t *OpenAI_ChatCompletionCreate(OpenAI_t *openai)
@@ -1191,6 +1319,7 @@ static OpenAI_ChatCompletion_t *OpenAI_ChatCompletionCreate(OpenAI_t *openai)
     _chatCompletion->parent.setUser = &OpenAI_ChatCompletionSetUser;
     _chatCompletion->parent.clearConversation = &OpenAI_ChatCompletionClearConversation;
     _chatCompletion->parent.message = &OpenAI_ChatCompletionMessage;
+    _chatCompletion->parent.multiModalMessage = &OpenAI_ChatCompletionMultiModalMessage;
 
     return &_chatCompletion->parent;
 }
@@ -1789,7 +1918,7 @@ static const char *audio_input_mime[] = {
     "audio/webm"
 };
 
-static const char *audio_speech_formats[] = {"mp3", "opus", "aac", "flac"};
+static const char *audio_speech_formats[] = {"mp3", "opus", "aac", "flac", "wav", "pcm"};
 
 /**
  * @brief Gives audio from the input text.
@@ -1852,7 +1981,7 @@ static void OpenAI_AudioSpeechSetSpeed(OpenAI_AudioSpeech_t *speech, float t)
 static void OpenAI_AudioSpeechSetResponseFormat(OpenAI_AudioSpeech_t *audioCreateSpeech, OpenAI_Audio_Output_Format rf)
 {
     _OpenAI_AudioSpeech_t *_audioCreateSpeech = __containerof(audioCreateSpeech, _OpenAI_AudioSpeech_t, parent);
-    if (rf >= OPENAI_AUDIO_OUTPUT_FORMAT_MP3 && rf <= OPENAI_AUDIO_OUTPUT_FORMAT_FLAC) {
+    if (rf >= OPENAI_AUDIO_OUTPUT_FORMAT_MP3 && rf < OPENAI_AUDIO_OUTPUT_FORMAT_MAX) {
         _audioCreateSpeech->response_format = rf;
     }
 }
@@ -1944,6 +2073,49 @@ OpenAI_SpeechResponse_t *OpenAI_AudioSpeechMessage(OpenAI_AudioSpeech_t *audioSp
     return result;
 }
 
+void OpenAI_AudioSpeechMessageStream(OpenAI_AudioSpeech_t *audioSpeech, char *p, OpenAI_StreamCallback stream_callback)
+{
+    size_t dataLength = 0;
+    const char *endpoint = "audio/speech";
+    cJSON *req = cJSON_CreateObject();
+    OPENAI_ERROR_CHECK_RETURN_VOID(req, "cJSON_CreateObject failed!");
+    _OpenAI_AudioSpeech_t *_audioSpeech = __containerof(audioSpeech, _OpenAI_AudioSpeech_t, parent);
+    if (cJSON_AddStringToObject(req, "model", (_audioSpeech->model == NULL) ? "tts-1" : _audioSpeech->model) == NULL) {
+        goto cleanup;
+    }
+    if (cJSON_AddStringToObject(req, "input", p) == NULL) {
+        goto cleanup;
+    }
+    if (cJSON_AddStringToObject(req, "voice", (_audioSpeech->voice == NULL) ? "alloy" : _audioSpeech->voice) == NULL) {
+        goto cleanup;
+    }
+    if (_audioSpeech->response_format != OPENAI_AUDIO_OUTPUT_FORMAT_MP3) {
+        if (cJSON_AddStringToObject(req, "response_format", audio_speech_formats[_audioSpeech->response_format]) == NULL) {
+            goto cleanup;
+        }
+    }
+    if (_audioSpeech->speed != 1.0) {
+        if (cJSON_AddNumberToObject(req, "speed", _audioSpeech->speed) == NULL) {
+            goto cleanup;
+        }
+    }
+    char *jsonBody = cJSON_Print(req);
+    ESP_LOGD(TAG, "json body for Speech Message %s", jsonBody);
+    cJSON_Delete(req);
+    req = NULL;
+
+    // the stream function will return NULL, so we don't need to free the result.
+    _audioSpeech->oai->speechpost_stream(_audioSpeech->oai->base_url, _audioSpeech->oai->api_key, endpoint, jsonBody, &dataLength, stream_callback);
+    free(jsonBody);
+    return;
+
+cleanup:
+    if (req) {
+        cJSON_Delete(req);
+    }
+    ESP_LOGE(TAG, "Failed to create json body for speech stream message");
+}
+
 static OpenAI_AudioSpeech_t *OpenAI_AudioSpeechCreate(OpenAI_t *openai)
 {
     _OpenAI_AudioSpeech_t *_audioCreateSpeech = (_OpenAI_AudioSpeech_t *)calloc(1, sizeof(_OpenAI_AudioSpeech_t));
@@ -1956,6 +2128,7 @@ static OpenAI_AudioSpeech_t *OpenAI_AudioSpeechCreate(OpenAI_t *openai)
     _audioCreateSpeech->parent.setSpeed = &OpenAI_AudioSpeechSetSpeed;
     _audioCreateSpeech->parent.setResponseFormat = &OpenAI_AudioSpeechSetResponseFormat;
     _audioCreateSpeech->parent.speech = &OpenAI_AudioSpeechMessage;
+    _audioCreateSpeech->parent.speechStream = &OpenAI_AudioSpeechMessageStream;
 
     return &_audioCreateSpeech->parent;
 }
@@ -2445,6 +2618,85 @@ end:
     return result != NULL ? result : NULL;
 }
 
+static char *OpenAI_Speech_Post(const char *base_url, const char *api_key, const char *endpoint, char *jsonBody, size_t *output_len)
+{
+    return OpenAI_Request(base_url, api_key, endpoint, "application/json", HTTP_METHOD_POST, NULL, (uint8_t *)jsonBody, strlen(jsonBody), output_len);
+}
+
+static char *OpenAI_Speech_Request_Stream(const char *base_url, const char *api_key, const char *endpoint, const char *content_type, esp_http_client_method_t method, const char *boundary, uint8_t *data, size_t len, size_t *output_len, OpenAI_StreamCallback stream_callback)
+{
+    ESP_LOGD(TAG, "\"%s\", len=%u", endpoint, len);
+    char *url = NULL;
+    asprintf(&url, "%s%s", base_url, endpoint);
+    OPENAI_ERROR_CHECK(url != NULL, "Failed to allocate url!", NULL);
+    esp_http_client_config_t config = {
+        .url = url,
+        .method = method,
+        .timeout_ms = 60000,
+        .crt_bundle_attach = esp_crt_bundle_attach,
+    };
+    esp_http_client_handle_t client = esp_http_client_init(&config);
+    char *headers = NULL;
+    if (boundary) {
+        asprintf(&headers, "%s; boundary=%s", content_type, boundary);
+    } else {
+        asprintf(&headers, "%s", content_type);
+    }
+    OPENAI_ERROR_CHECK_GOTO(headers != NULL, "Failed to allocate headers!", end);
+    esp_http_client_set_header(client, "Content-Type", headers);
+    ESP_LOGD(TAG, "headers:\r\n%s", headers);
+    free(headers);
+
+    asprintf(&headers, "Bearer %s", api_key);
+    OPENAI_ERROR_CHECK_GOTO(headers != NULL, "Failed to allocate headers!", end);
+    esp_http_client_set_header(client, "Authorization", headers);
+    free(headers);
+
+    esp_err_t err = esp_http_client_open(client, len);
+    ESP_LOGD(TAG, "data:\r\n%s", data);
+
+    OPENAI_ERROR_CHECK_GOTO(err == ESP_OK, "Failed to open client!", end);
+    if (len > 0) {
+        int wlen = esp_http_client_write(client, (const char *)data, len);
+        OPENAI_ERROR_CHECK_GOTO(wlen >= 0, "Failed to write client!", end);
+    }
+    int content_length = esp_http_client_fetch_headers(client);
+    if (esp_http_client_is_chunked_response(client)) {
+        esp_http_client_get_chunk_length(client, &content_length);
+    }
+    ESP_LOGD(TAG, "chunk_length=%d", content_length); //4096
+    OPENAI_ERROR_CHECK_GOTO(content_length > 0, "HTTP client fetch headers failed!", end);
+
+    int read_len = 0;
+    *output_len = 0;
+    const uint32_t chunk_size = 1024 * 33;
+    uint8_t * chunk_data = (uint8_t *)malloc(chunk_size);
+    if (!chunk_data) {
+        ESP_LOGE(TAG, "Failed to allocate chunk_data");
+        goto end;
+    }
+    do {
+        read_len = esp_http_client_read_response(client, (char*)chunk_data, chunk_size);
+        if (stream_callback) {
+            stream_callback(chunk_data, read_len);
+        }
+        *output_len += read_len;
+        ESP_LOGD(TAG, "HTTP_READ:=%d", read_len);
+    } while (read_len > 0);
+    ESP_LOGD(TAG, "output_len: %d\n", (int)*output_len);
+    free(chunk_data);
+end:
+    free(url);
+    esp_http_client_close(client);
+    esp_http_client_cleanup(client);
+    return NULL;
+}
+
+static char *OpenAI_Speech_Post_Stream(const char *base_url, const char *api_key, const char *endpoint, char *jsonBody, size_t *output_len, OpenAI_StreamCallback cb)
+{
+    return OpenAI_Speech_Request_Stream(base_url, api_key, endpoint, "application/json", HTTP_METHOD_POST, NULL, (uint8_t *)jsonBody, strlen(jsonBody), output_len, cb);
+}
+
 static char *OpenAI_Upload(const char *base_url, const char *api_key, const char *endpoint, const char *boundary, uint8_t *data, size_t len)
 {
     return OpenAI_Request(base_url, api_key, endpoint, "multipart/form-data", HTTP_METHOD_POST, boundary, data, len, NULL);
@@ -2529,6 +2781,8 @@ OpenAI_t *OpenAICreate(const char *api_key)
     _oai->get = &OpenAI_Get;
     _oai->del = &OpenAI_Del;
     _oai->post = &OpenAI_Post;
+    _oai->speechpost = &OpenAI_Speech_Post;
+    _oai->speechpost_stream = &OpenAI_Speech_Post_Stream;
     _oai->upload = &OpenAI_Upload;
     return &_oai->parent;
 }